% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tetris.R
\name{tetris}
\alias{tetris}
\title{Tetris for R}
\usage{
tetris(length, width, data = rep(0, length * width))
}
\arguments{
\item{length}{length of tetris board}

\item{width}{width of tetris board}

\item{data}{current position of pieces on tetris board}
}
\value{
plays tetris
}
\description{
Tetris for R
}

